example_problem_test.py
**Two files: 
the program(example_problem.py) and the test for the program (example_problem_test.py)

Exercise instructions: Create a simple tip calculator. The program should prompt for a bill amount and a tip rate. The program must compute the tip and then display both the tip and the total amount of the bill.