flower_reader.py
Instructions: Create a function that opens the flowers.txt, reads every line in it, and saves it as a dictionary. The main (separate) function should take user input (user's first name and last name) and parse the user input to identify the first letter of the first name. It should then use it to print the flower name with the same first letter (from dictionary created in the first function).

Ex: >>> Enter your First [space] Last name only: Bill Newman
>>> Unique flower name with the first letter: Bellflower